requirements.txt
django-notifications-hq==1.6.0
Django==2.2.12
django-crispy-forms==1.7.2
django-extensions==2.2.8
django-filter==2.2.0
html5lib==1.0.1
Pillow==6.1.0
PyPDF2==1.26.0
pytz==2018.9
reportlab==3.5.19
six==1.12.0
webencodings==0.5.1
xhtml2pdf==0.2.3
